Fix FileProvider conflicts
The build of a Tabris.js app that contains the 'tabris-plugin-camera' plugin fails due to the manifets merger exception. The exception occurs because of the FileProvider, which specified in the manifets file in both in the plugin and the 'tabris-android' platform. The same exception will be thrown again if another plugin also specifies the same provider.

The change creates a custom FileProvider, which just extends FileProvider and specifies in the manifest to avoid such conflicts.